realized seed file existed in scripts - moved seed data over and deleted other file
